docs: update readme with content from docs.getunleash.io (#136)
## What

This PR adds content that was available at docs.getunleash.io to the readme. It also restructures the readme a little based on the incoming content.
## Why

Because we're moving to automating SDK docs on docs.getunleash.io and we don't want to lose any content. Some of the content on docs.getunleash.io was missing or more informative, so I've incorporated it as best as I can.

## commits

* Docs: merge content from docs.getunleash.io

* docs: prioritize front-and API over proxy