chore(1-3230): use homebrew version of uuid generation (#240)
The one from the uuid library relies on an underlying crypto library
which doesn't exist at least in certain GitHub runners and may also
cause issues for react native applications.

This impl sidesteps that issue.


The same uuid generation method that we're cutting out here is also being used in src/events-handler.ts. However, because we haven't received any complaints about that, I'll leave it in.